+++ Backend-Ecommerce/src/services/models/modelCart.ts
@@ -4,6 +4,7 @@ import schemaCart from '../schemas/schemaCart';
 export interface IProducts{
     product_id: ObjectId;
     quantity: number;
+    colors: [string];
 }
 
 export interface ICart extends Document{