+++ Backend-Ecommerce/src/services/schemas/schemaCart.ts
@@ -5,7 +5,8 @@ const schemaCart = new Schema({
     user_id: {type: Schema.Types.ObjectId, ref: 'User', required: true},
     products:[{
         product_id:{type: Schema.Types.ObjectId, ref: 'Products', required: true},
-        quantity: {type: Number, required:true}
+        quantity: {type: Number, required:true},
+        colors: {type: [String], required: true}
     }],
     datecreated: {type: Date, default: Date.now}
 })